requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
python-dotenv==1.0.0
email-validator==2.0.0
